add some static_cast to remove the warning.

+++ libproject/Utility/HTTPRequestPacket.cpp
@@ -36,19 +36,19 @@ HTTPRequestPacket::~HTTPRequestPacket(void) {
 
 int HTTPRequestPacket::getMainHostName(char *buffer, const int len) {
 	strncpy(buffer, main_host_, getWrittenCount(len));
-	return strlen(buffer);
+	return static_cast<int>(strlen(buffer));
 }
 int HTTPRequestPacket::getHost(char *buffer, const int len) {
 	strncpy(buffer, host_, getWrittenCount(len));
-	return strlen(buffer);
+	return static_cast<int>(strlen(buffer));
 }
 int HTTPRequestPacket::getReferer(char *buffer, const int len) {
 	strncpy(buffer, referer_, getWrittenCount(len));
-	return strlen(buffer);
+	return static_cast<int>(strlen(buffer));
 }
 int HTTPRequestPacket::getUserAgent(char *buffer, const int len) {
 	strncpy(buffer, useagent_, getWrittenCount(len));
-	return strlen(buffer);
+	return static_cast<int>(strlen(buffer));
 }
 
 void HTTPRequestPacket::reset() {
@@ -73,7 +73,7 @@ int HTTPRequestPacket::parsePacket(WSABUF * wsabuf, const int count) {
 int HTTPRequestPacket::getGET(char *buffer, const int len) {
 	if (HTTP_REQUEST_OPETYPE_GET == getRequestType()) {
 		strncpy(buffer, oper_, len);
-		return strlen(buffer);
+		return static_cast<int>(strlen(buffer));
 	}
 
 	return 0;
@@ -82,7 +82,7 @@ int HTTPRequestPacket::parsePacket(char * buf, const int len) {
 	using namespace strutility;
 
 	const char * HTTP_REQUEST_NEW_LINE = "\r\n";
-	const int  HTTP_REQUEST_NEW_LINE_LENGTH = strlen(HTTP_REQUEST_NEW_LINE);
+	const int  HTTP_REQUEST_NEW_LINE_LENGTH = static_cast<int>(strlen(HTTP_REQUEST_NEW_LINE));
 	//reset buffer 
 	reset();
 
@@ -94,26 +94,26 @@ int HTTPRequestPacket::parsePacket(char * buf, const int len) {
 	while (next != NULL) {
 		if (strstr(p, HTTP_REQUEST_HOST) == p) {
 			// 主机地址只留DNS主名
-			const int tab_length = strlen(HTTP_REQUEST_HOST);
+			const int tab_length = static_cast<int>(strlen(HTTP_REQUEST_HOST));
 			strncpy(host_, p + tab_length, getWrittenCount(next - p - tab_length));
 			get_main_dns_name(main_host_, HTTP_REQUEST_ITEM_MAX_LENGTH, host_);
 			cnt++;
 		} else if (strstr(p, HTTP_REQUEST_REFERER) == p) {
-			const int tab_length = strlen(HTTP_REQUEST_REFERER);
+			const int tab_length = static_cast<int>(strlen(HTTP_REQUEST_REFERER));
 			strncpy(referer_, p + tab_length, getWrittenCount(next - p - tab_length));
 			cnt++;
 		} else if (strstr(p, HTTP_REQUEST_OPER_GET) == p) {
-			const int tab_length = strlen(HTTP_REQUEST_OPER_GET);
+			const int tab_length = static_cast<int>(strlen(HTTP_REQUEST_OPER_GET));
 			strncpy(oper_, p + tab_length, getWrittenCount(next - p - tab_length));
 			http_request_type_ = HTTP_REQUEST_OPETYPE_GET;
 			cnt++;
 		} else if (strstr(p, HTTP_REQUEST_OPER_POST) == p) {
-			const int tab_length = strlen(HTTP_REQUEST_OPER_POST);
+			const int tab_length = static_cast<int>(strlen(HTTP_REQUEST_OPER_POST));
 			strncpy(oper_, p + tab_length, getWrittenCount(next - p - tab_length));
 			http_request_type_ = HTTP_REQUEST_OPETYPE_POST;
 			cnt++;
 		} else if (strstr(p, HTTP_REQUEST_USER_AGENT) == p) {
-			const int tab_length = strlen(HTTP_REQUEST_USER_AGENT);
+			const int tab_length = static_cast<int>(strlen(HTTP_REQUEST_USER_AGENT));
 			strncpy(useagent_, p + tab_length, getWrittenCount(next-p-tab_length));
 			cnt++;
 		} else if (strstr(p, HTTP_REQUEST_OPER_POST) == p) {
